requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
discord.py==2.3.4
dropbox==13.8.0